Fix GH#23652: Possible buffer overflow in `Capella::readString()`

+++ src/importexport/capella/internal/capella.h
@@ -685,7 +685,7 @@ struct CapBracket {
 class Capella
 {
     static const char* errmsg[];
-    int curPos;
+    qint64 curPos;
 
     QFile* f;
     char* author;
